Add Dexie schema (#46)
Because Dexie is exposed as a module, in order
to bring it in properly, we need a bundler of
some kind so we can have the module imported.

Create schema and also add Webpack config
to allow for bundling of Dexie in our project. Add
"build" NPM task to account for Webpack build.

In order to import database wrapper,
import database object from 'dist/depBundle.js'.
Make sure to build first. Additionally, read Dexie
docs for details on how to use it.

https://dexie.org/docs/Tutorial/Getting-started